requirements.txt
fastapi
uvicorn
requests
python-dotenv
msal
snowflake-connector-python
pydantic
groq